+++ ballast-core/src/commonMain/kotlin/com/copperleaf/ballast/Queued.kt
@@ -1,5 +1,7 @@
 package com.copperleaf.ballast
 
+import kotlinx.coroutines.CompletableDeferred
+
 public sealed class Queued<Inputs : Any, Events : Any, State : Any> {
 
     public class RestoreState<Inputs : Any, Events : Any, State : Any>(
@@ -7,6 +9,7 @@ public sealed class Queued<Inputs : Any, Events : Any, State : Any> {
     ) : Queued<Inputs, Events, State>()
 
     public class HandleInput<Inputs : Any, Events : Any, State : Any>(
+        public val deferred: CompletableDeferred<Unit>?,
         public val input: Inputs,
     ) : Queued<Inputs, Events, State>()
 }