+++ ballast-core/src/commonMain/kotlin/com/copperleaf/ballast/internal/EventHandlerScopeImpl.kt
@@ -2,14 +2,13 @@ package com.copperleaf.ballast.internal
 
 import com.copperleaf.ballast.BallastLogger
 import com.copperleaf.ballast.EventHandlerScope
-import kotlinx.coroutines.channels.SendChannel
 
 internal class EventHandlerScopeImpl<Inputs : Any, Events : Any, State : Any>(
     override val logger: BallastLogger,
-    private val _inputs: SendChannel<Inputs>,
+    private val sendInputToViewModel: suspend (Inputs)->Unit,
 ) : EventHandlerScope<Inputs, Events, State> {
     override suspend fun postInput(input: Inputs) {
-        _inputs.send(input)
+        sendInputToViewModel(input)
     }
 
     fun ensureUsedCorrectly() {