+++ ballast-test/src/commonMain/kotlin/com/copperleaf/ballast/test/internal/vm/ViewModelWrapper.kt
@@ -2,11 +2,8 @@ package com.copperleaf.ballast.test.internal.vm
 
 import com.copperleaf.ballast.BallastViewModel
 import kotlinx.coroutines.CompletableDeferred
-import kotlinx.coroutines.ExperimentalCoroutinesApi
 import kotlinx.coroutines.channels.ChannelResult
-import kotlinx.coroutines.channels.SendChannel
 import kotlinx.coroutines.flow.StateFlow
-import kotlinx.coroutines.selects.SelectClause2
 
 internal class ViewModelWrapper<Inputs : Any, Events : Any, State : Any>(
     private val delegate: BallastViewModel<TestViewModel.Inputs<Inputs>, Events, State>,
@@ -17,24 +14,12 @@ internal class ViewModelWrapper<Inputs : Any, Events : Any, State : Any>(
     override val type: String
         get() = delegate.type
 
-    @ExperimentalCoroutinesApi
-    override val isClosedForSend: Boolean
-        get() = delegate.isClosedForSend
-
-    override val onSend: SelectClause2<Inputs, SendChannel<Inputs>>
-        get() = error("cannot call onSend on ViewModelWrapper")
-
-    override fun close(cause: Throwable?): Boolean {
-        return delegate.close(cause)
-    }
-
-    @ExperimentalCoroutinesApi
-    override fun invokeOnClose(handler: (cause: Throwable?) -> Unit) {
-        delegate.invokeOnClose(handler)
-    }
-
     override suspend fun send(element: Inputs) {
         delegate.send(TestViewModel.Inputs.ProcessInput(element, CompletableDeferred()))
+    }
+
+    override suspend fun sendAndAwaitCompletion(element: Inputs) {
+        delegate.sendAndAwaitCompletion(TestViewModel.Inputs.ProcessInput(element, CompletableDeferred()))
     }
 
     override fun trySend(element: Inputs): ChannelResult<Unit> {
@@ -44,4 +29,5 @@ internal class ViewModelWrapper<Inputs : Any, Events : Any, State : Any>(
     override fun observeStates(): StateFlow<State> {
         return delegate.observeStates()
     }
+
 }