+++ CHANGELOG.md
@@ -8,6 +8,8 @@
   - By writing a custom `InputStrategy.Guardian`, you could remove the restriction that sideJobs must be started at the
     end of the Input, and potentially actually call `sideJob()` and wait for the sideJob to start before continuing. 
     This is dangerous and should not be taken lightly, but it is now possible, whereas it was not before.
+- `BallastViewModel` no longer implements the `SendChannel` interface, it exposes the `send` and `trySend` methods 
+  directly. It also adds `sendAndAwaitCompletion()` to send an Input and wait for it to completely finish processing.
 
 ## 0.12.0 - 2022-03-17
 
@@ -39,7 +41,7 @@
 - Adds a few new `BallastNotification`s, to notify of Inputs and Events being queued, but not yet processed
 - Fixed bug where SideEffect errors and cancellation was not being notified properly
 - `SideEffectcope` is now also a `CoroutineScope`, so you can `launch` other coroutines directly within the 
-  `sideJob(key) { }` block without having to manually wrap it in `coroutineScope { }`
+  `sideEffect(key) { }` block without having to manually wrap it in `coroutineScope { }`
 
 ## 0.8.2 - 2022-02-23
 
@@ -56,7 +58,7 @@ changes, as descibed below:
 
 - Improves side-effects
   - They are now started through a Channel, instead of relying on a Mutex for safety
-  - New `BallastInterceptor` methods added to track sideJob starts/errors
+  - New `BallastInterceptor` methods added to track sideEffect starts/errors
   - SideEffects (and related APIs) now _require_ a `key` to be provided, to avoid accidental restarting cause by 
     disparate Inputs forgetting to provide a key
 - Improves safety of `ParallelInputStrategy` to prevent multiple updates, rather than just throwing an error at the end